Add cachinf for database queries
This commit adds a (barebone) cache class that allows us to store database queries results for a certain amount of time.
This will help to make geofencing queries faster.
For now, we will keep up to 1000 items in cache for 24h.
The decision to write a specific implementation instead of using a third-party dependency was motivated by the lack of advanced caching libraries in the js/ts ecosystem. So, right now we have something that is up to the standards of other libraries and we will actually be able to modify and scale it to our own needs.