Remove ElysiaJS
Bun now provides its own HTTP router, which was basically the only reason we were using ElysiaJS, which came with design constraints.
By removing this dependency, we can now write our code as close to the core engine as we can, which will allow us to make our design evolve in whatever direction we'll need.
Writing small utils functions is a very small price to pay for this kind of freedom.
This rewriting already simplifies the process of adding namespaces to the API, and we will surely be able to simplify it even more in the future.